[TD-5623]<feature>: fixed crash caused by second dnodeFetchQ thread try access pRuntimeEnv->outputbuf NULL pointer issue

+++ src/query/src/queryMain.c
@@ -338,7 +338,7 @@ int32_t qDumpRetrieveResult(qinfo_t qinfo, SRetrieveTableRsp **pRsp, int32_t *co
   }
 
   (*pRsp)->precision = htons(pQueryAttr->precision);
-  (*pRsp)->compressed = checkNeedToCompressQueryCol(pQInfo);
+  (*pRsp)->compressed = (int8_t)checkNeedToCompressQueryCol(pQInfo);
 
   if (GET_NUM_OF_RESULTS(&(pQInfo->runtimeEnv)) > 0 && pQInfo->code == TSDB_CODE_SUCCESS) {
     doDumpQueryResult(pQInfo, (*pRsp)->data, (*pRsp)->compressed, &compLen);
